feat: release v2.2.0 with updated bundle path and minor improvements
- Moved SupervisorBundle to Infrastructure\Symfony namespace
- Updated README with deployment recommendation
- Added strict type annotations to improve PHPStan compatibility
