+++ .gitignore
@@ -1,1 +1,5 @@
 .idea
+/vendor/
+/composer.lock
+/.php-cs-fixer.cache
+/.phpunit.result.cache
