+++ docker-compose.yaml
@@ -1,0 +1,7 @@
+services:
+  php:
+    build: .
+    volumes:
+      - .:/app
+    working_dir: /app
+    command: [ "php", "-a" ]
