+++ src/Infrastructure/Symfony/SupervisorBundle.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pierotto\SupervisorBundle\Infrastructure\Symfony;
+
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+class SupervisorBundle extends Bundle
+{
+}
