--- src/Resources/config/services.yml
@@ -1,12 +1,0 @@
-services:
-
-    _defaults:
-        autowire: TRUE
-        autoconfigure: TRUE
-
-    Pierotto\SupervisorBundle\Application\Command\SupervisorGeneratorCommand:
-
-    Pierotto\SupervisorBundle\Domain\SupervisorConfigGenerator:
-        arguments:
-            $prefix: '%supervisor.prefix%'
-            $programs: '%supervisor.programs%'
