--- src/SupervisorBundle.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types = 1);
-
-namespace Pierotto\SupervisorBundle;
-
-class SupervisorBundle extends \Symfony\Component\HttpKernel\Bundle\Bundle
-{
-
-}
